CrossyC: option to install binutils only

+++ Formula/crossyc.rb
@@ -5,12 +5,14 @@ class Crossyc < Formula
   mirror "https://ftp.gnu.org/gnu/gcc/gcc-6.1.0/gcc-6.1.0.tar.bz2"
   sha256 "09c4c85cabebb971b1de732a0219609f93fc0af5f86f6e437fd8d7f832f1a351"
 
-  option "without-c++", "Build without the c++ compiler"
+  option "without-gcc", "Build binutils only (no gcc compiler)"
 
-  depends_on "gmp"
-  depends_on "libmpc"
-  depends_on "mpfr"
-  depends_on "isl"
+  if ! build.without? "gcc" do
+    depends_on "gmp"
+    depends_on "libmpc"
+    depends_on "mpfr"
+    depends_on "isl"
+  end
 
   resource "binutils" do
     url "https://ftpmirror.gnu.org/binutils/binutils-2.26.1.tar.gz"
@@ -40,34 +42,33 @@ class Crossyc < Formula
       info.rmtree
     end
 
-    ENV["PATH"] += ":#{bin}"
+    if ! build.without? "gcc" do
+      ENV["PATH"] += ":#{bin}"
 
-    languages = %w[c]
-    languages << "c++" unless build.without? "c++"
+      args = [
+        "--target=x86_64-elf",
+        "--prefix=#{prefix}",
+        "--enable-languages=c,c++",
+        "--with-gmp=#{Formula["gmp"].opt_prefix}",
+        "--with-mpfr=#{Formula["mpfr"].opt_prefix}",
+        "--with-mpc=#{Formula["libmpc"].opt_prefix}",
+        "--with-isl=#{Formula["isl"].opt_prefix}",
+        "--without-headers",
+        "--disable-nls",
+        "--disable-werror",
+      ]
 
-    args = [
-      "--target=x86_64-elf",
-      "--prefix=#{prefix}",
-      "--enable-languages=#{languages.join(",")}",
-      "--with-gmp=#{Formula["gmp"].opt_prefix}",
-      "--with-mpfr=#{Formula["mpfr"].opt_prefix}",
-      "--with-mpc=#{Formula["libmpc"].opt_prefix}",
-      "--with-isl=#{Formula["isl"].opt_prefix}",
-      "--without-headers",
-      "--disable-nls",
-      "--disable-werror",
-    ]
+      mkdir "build" do
+        system "../configure", *args
+        system "make", "all-gcc"
+        system "make", "all-target-libgcc"
 
-    mkdir "build" do
-      system "../configure", *args
-      system "make", "all-gcc"
-      system "make", "all-target-libgcc"
+        system "make", "install-gcc"
+        system "make", "install-target-libgcc"
+      end
 
-      system "make", "install-gcc"
-      system "make", "install-target-libgcc"
+      man7.rmtree
+      info.rmtree
     end
-
-    man7.rmtree
-    info.rmtree
   end
 end
